backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
billiard==4.2.2
celery==5.4.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.0
gotrue==1.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
kombu==5.6.0
multidict==6.7.0
packaging==25.0
pillow==10.4.0
postgrest==2.24.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.9
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.24.0
redis==5.0.8
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
supafunc==0.3.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
